Fixed botBattle by using push_san instead of loading board everytime

+++ src/base/kings.rs
@@ -66,7 +66,7 @@ impl Board {
             // Remove castling bits as King is moved
             new_board.remove_castling_bits(CastleSide::King, &curr_colour);
             new_board.remove_castling_bits(CastleSide::Queen, &curr_colour);
-            new_board.latest_move = ( ( index << 6 ) as u8 | new_index) as u32;
+            new_board.latest_move = ( (index as u32) << 6 | new_index as u32) as u32;
             new_board.latest_move &= ( 1 << 12 ) - 1;
             legal_boards.push(&mut new_board);
         }
