+++ src/base/pawns.rs
@@ -82,7 +82,7 @@ impl Board {
                 new_board.pawns |= 1 << 64 * is_black + new_pos; // Update new pawn position
                 new_board.update_tickers(true, is_black == 1); // Update Tickers
                 new_board.set_enpassant( None );
-                new_board.latest_move = ( ( ( index as u8 ) << 6 ) | (new_index as u8) ) as u32;
+                new_board.latest_move = ( (index as u32) << 6 | new_index as u32) as u32;
                 new_board.latest_move &= ( 1 << 12 ) - 1;
                 new_board.check_and_add_promotion(new_index, is_black, legal_boards);
             }
@@ -101,7 +101,7 @@ impl Board {
                     new_board.pawns |= 1 << 64 * is_black + new_pos; // Update new pawn position
                     new_board.update_tickers(true, is_black == 1); // Update Tickers
                     new_board.set_enpassant( Some( x as u8 ) ); // mark en-passant possible at current x.
-                    new_board.latest_move = ( ( index << 6 ) as u8 | new_index) as u32;
+                    new_board.latest_move = ( (index as u32) << 6 | new_index as u32) as u32;
                     new_board.latest_move &= ( 1 << 12 ) - 1;
                     legal_boards.push(&mut new_board);
                 }
@@ -124,7 +124,7 @@ impl Board {
                     new_board.pawns |= 1 << 64 * is_black + new_pos; // Update new pawn position
                     new_board.update_tickers(true, is_black == 1); // Update Tickers
                     new_board.set_enpassant( None );
-                    new_board.latest_move = ( ( ( index as u8 ) << 6 ) | (new_index as u8) ) as u32;
+                    new_board.latest_move = ( (index as u32) << 6 | new_index as u32) as u32;
                     new_board.latest_move &= ( 1 << 12 ) - 1;
                     new_board.check_and_add_promotion(new_index, is_black, legal_boards);
                 }
@@ -149,7 +149,7 @@ impl Board {
 
                     new_board.update_tickers(true, is_black == 1); // Update Tickers
                     new_board.set_enpassant( None );
-                    new_board.latest_move = ( ( ( index as u8 ) << 6 ) | (new_index as u8) ) as u32;
+                    new_board.latest_move = ( (index as u32) << 6 | new_index as u32) as u32;
                     new_board.latest_move &= ( 1 << 12 ) - 1;
                     legal_boards.push(&mut new_board);
                 }
